# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "AcademiaConnect API",
                 "description":  "Backend API for AcademiaConnect SIH 2026",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/student/profile/{user_id}":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "Student Profile"
                                                                          ],
                                                                 "summary":  "Get Student Profile",
                                                                 "operationId":  "get_student_profile_student_profile__user_id__get",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "user_id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "User Id"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {

                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             }
                                                 },
                  "/student/profile/":  {
                                            "put":  {
                                                        "tags":  [
                                                                     "Student Profile"
                                                                 ],
                                                        "summary":  "Update Student Profile",
                                                        "operationId":  "update_student_profile_student_profile__put",
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/StudentProfileRequest"
                                                                                                                                }
                                                                                                                 }
                                                                                        },
                                                                            "required":  true
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {

                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  }
                                                                      }
                                                    }
                                        },
...
                  "/dashboard/company/{user_id}":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "Dashboard"
                                                                            ],
                                                                   "summary":  "Company Dashboard",
                                                                   "operationId":  "company_dashboard_dashboard_company__user_id__get",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "user_id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "title":  "User Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {

                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                   },
...
                  "/company/profile/":  {
                                            "put":  {
                                                        "tags":  [
                                                                     "Company Profile"
                                                                 ],
                                                        "summary":  "Update Company Profile",
                                                        "operationId":  "update_company_profile_company_profile__put",
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/CompanyProfileRequest"
                                                                                                                                }
                                                                                                                 }
                                                                                        },
                                                                            "required":  true
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {

                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  }
                                                                      }
                                                    }
                                        },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CompanyProfileRequest":  {
                                                                     "properties":  {
                                                                                        "user_id":  {
                                                                                                        "type":  "integer",
                                                                                                        "title":  "User Id"
                                                                                                    },
                                                                                        "name":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Name",
                                                                                                     "default":  ""
                                                                                                 },
                                                                                        "industry":  {
                                                                                                         "type":  "string",
                                                                                                         "title":  "Industry",
                                                                                                         "default":  ""
                                                                                                     },
                                                                                        "description":  {
                                                                                                            "type":  "string",
                                                                                                            "title":  "Description",
                                                                                                            "default":  ""
                                                                                                        },
                                                                                        "location":  {
                                                                                                         "type":  "string",
                                                                                                         "title":  "Location",
                                                                                                         "default":  ""
                                                                                                     }
                                                                                    },
                                                                     "type":  "object",
                                                                     "required":  [
                                                                                      "user_id"
                                                                                  ],
                                                                     "title":  "CompanyProfileRequest"
                                                                 },
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "StudentProfileRequest":  {
                                                                     "properties":  {
                                                                                        "user_id":  {
                                                                                                        "type":  "integer",
                                                                                                        "title":  "User Id"
                                                                                                    },
                                                                                        "college":  {
                                                                                                        "type":  "string",
                                                                                                        "title":  "College",
                                                                                                        "default":  ""
                                                                                                    },
                                                                                        "branch":  {
                                                                                                       "type":  "string",
                                                                                                       "title":  "Branch",
                                                                                                       "default":  ""
                                                                                                   },
                                                                                        "semester":  {
                                                                                                         "type":  "string",
                                                                                                         "title":  "Semester",
                                                                                                         "default":  ""
                                                                                                     },
                                                                                        "cgpa":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Cgpa",
                                                                                                     "default":  ""
                                                                                                 },
                                                                                        "career_goal":  {
                                                                                                            "type":  "string",
                                                                                                            "title":  "Career Goal",
                                                                                                            "default":  ""
                                                                                                        }
                                                                                    },
                                                                     "type":  "object",
                                                                     "required":  [
                                                                                      "user_id"
                                                                                  ],
                                                                     "title":  "StudentProfileRequest"
                                                                 },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           },
...
                                   }
                   }
}